topic_modeling/requirements.txt
Django==2.0.12
djangorestframework==3.9.4
djongo==1.2.32
pymongo==3.8.0
numpy==1.16.3
pandas==0.24.2
scipy==1.3.0
scikit-learn==0.21.1
pyLDAvis==2.1.2
nltk==3.4.1
Unipath==1.1